Conflicts:
	Makefile

+++ include/linux/soc/qcom/qmi.h
@@ -167,7 +167,7 @@ struct qmi_ops {
 struct qmi_txn {
 	struct qmi_handle *qmi;
 
-	int id;
+	u16 id;
 
 	struct mutex lock;
 	struct completion completion;
